Refactored configuration retrieval
Settings don't have to be published anymore because they no longer need to be linked to the root pages of the sites. Instead the Settings are now found by convention. They now have to reside in specific folders to be found.
This should further reduce the risk of leaking configuration through customizable clients that have access to the repository.