fix: Work around GCC Java REST Client Issue
While the GCC REST API documents, that times like due dates should
be given in UTC, the Java REST Client API uses the local time-zone
instead for validating incoming due dates.

This caused an IllegalArgumentException for tests, that just use an
offset of some hours from the current time in time-zones such as
`America/Chicago`.

The main fix is to adjust the offset for a due date to something more
typical, i.e., some days rather than some hours ahead.

In addition to that, extended the corresponding tests, to validate
with different time-zones.

fixes: #59