test: Adapted Cancelation Failure Test
As it seems, derived from the existing test-steps, _Scenario:
Cancelation Failure Handling_ was meant to cause a failure just because
two workflows get canceled in parallel.

At least current observation shows, that this is not true. Instead,
as it is about testing cancelation and error handling behavior, added
a "Break it" step, where we manipulate the credentials, so that
subsequent communication will fail.

Also added a test for "multi-selection" within the workflow app, as
this is also relevant.