fix(workflow-server): Respect Redelivered State in Workflow
We now handle the Redelivered state as part of the translation
workflow. Due to its subtle difference to the `Delivered` state,
we introduce a new user task `ReviewRedeliveredTranslation` in parallel
to the existing user task `ReviewDeliveredTranslation`.

The difference is, at least in an observed scenario, that to process
the subsequent user-task it is not only required to "review" the
delivered translation, but also, to manually upload an XLIFF provided
by email, for example.

Thus, instead of reusing the existing `ReviewDeliveredTranslation` by
just handling `Redelivered` and `Delivered` in the same way, we want to
raise awareness for editors, that there may be something more to do by
this extra user-task.