feat(mock): Support Mocking Irregular Submission States
At least from the API perspective, states such as `Redelivered` are
considered irregular (and not even documented anywhere for the GCC
API).

To be still able to test this irregular state, we extend the mocking
configuration, to be able to override "normal" state sequences.
So, for example, we may now replace `DELIVERED` state to be handled
as `REDELIVERED` or we may intercept the `COMPLETED` state to continue
to the `DELIVERED` state.

The provided Mock API is quite rich regarding its possibilities and
should be handled with care, as you may provoke states, that you can
never reach via normal processing with GCC.