refactor: Make the SubmissionStore Singleton Instance Self-Contained
As the `MockGCExchangeFacade` is meant to be recreated several times
during a typical translation workflow execution, the `SubmissionStore`
must be a singleton. To ensure, we do not break this pattern during
refactorings, refactored the `SubmissionStore` to be a
"self-managed (lazy) singleton".

Also, in preparation of subsequent refactoring, replacing property
based (re-)configuration of the `SubmissionStore` to accepting the
(new) settings object instead.