fix: Check "Done" States First
Status such as `Delivered`, `Redelivered` and `Cancelled` (BE; sic!)
are all considered final states, where we do not expect to download
anymore.

To understand, it may be important to know, that from this client's
perspective the state `Delivered` is only reached after successful
download and XLIFF import, as only then the API is called to mark
corresponding tasks (and subsequently the submission) as `Done` aka
`Delivered`.

TL;DR - Moved all "final state options" before checking for possible
download translation errors.