feat: Widen XLIFF Download
Especially in context of "Redelivered" State, we also assume the XLIFF
artifact to contain information on possible errors. So, that is why
we widen the XLIFF download also for the "Redelivered" state.

Note, that if the Blob is empty (it, for example, is unavailable, if
no issues exist), the Download Button is also hidden.