fix: Fail Early on Invalid Connector Key
We observed, that the connector key is not or not always validated as
expected. Instead, for example when trying to retrieve a submission by
ID with an invalid connector key, it just returned "no submission found"
instead of complaining.

This caused, for example, endless retries when performing the manual
test for a failed cancelation communication. For the initial request
for a submission by ID, our facade instead returned some "mocked"
submission in state "OTHER", which subsequently is not handled by any
means within the workflow definition.