test: Extend Contract Test for Instructions
First test for instructions. We can, though, just check, that we can
create such a submission. We cannot validate it, as it is not exposed
within the REST API (at least on Java level).

So, the idea is, that subsequently we extend our manual tests to take
a look at these submissions having instructions.