fix!: Prevent Escalation for Higher Unicode Characters
Found that the GCC Backend has issues forwarding characters above
the Basic Multilingual Plane (BMP). Also, the GCC backend expects the
instructions to be given in HTML, which again breaks for workflow notes
that are typically provided as plain-text.

To prevent surprises, the default is now text-bmp, thus, we assume
incoming plain-text and we assume, that the GCC backend, as it is now
does not like higher level Unicode characters.

To get back to old behavior, you may set the type to `html` which is
similar to "take as it is". For robustness, we suggest using `html-bmp`
instead, as also for HTML higher Unicode characters are not supported.

BREAKING-CHANGE: Instruction Texts will now be modified. To get back to
previous behavior, set `html` as type.