fix(gcc-facade): Don't Overwrite PD Submission IDs
If we fail to retrieve a submission by its ID (for example, when
someone broke the GlobalLink settings during workflow execution) we
now do not overwrite the "PD Submission IDs" (PD = Project Director)
anymore.

This fixes an issue previously introduced when refactoring the
"Unknown Submission ID" representation from an artificial Submission
object to an exception, which means, that `getSubmission()` may now
escalate and prevent subsequent code to be executed.