feat(mock): Allow Mocking a Submission Error State
While for now we can trigger an error within the GCC backend using
instructions with characters from the Unicode Supplementary Multilingual
Plane, this may change over time.

To still cover the state, that a submission may reach an error state,
extending the mock facade to support a new error type
`submission_error`.

This makes it easier to test our workflow and our localization of this
state in Studio without relying on the actual GCC backend behavior.