fix(gcc-facade): Submission Name Character Types
Just as for the submission instructions, we identified, that also the
submission name must only contain characters from the Unicode Basic
Multilingual Plane.

In contrast to the submission instructions, where we considered it
valuable to keep a hint to a given character (by its Unicode Code
Point), we default to replacing problematic characters in the
submission name with underscores. This aligns with the behavior of
Project Director (PD), that also transforms other problematic
characters internally.