fix(gcc-facade): Document Content Name Character Limitation
Just as for the submission instructions, and submission names, we
identified, that also the name for uploaded contents must only contain
characters from the Unicode Basic Multilingual Plane (BMP).

Unlike for the submission instructions and submission name, we only
documented this limitation. In context of this connector, the
filenames are auto-generated from language tags that are not expected
to contain, for example, characters from the Unicode Supplementary
Plane.