docs: Intermediate Alternative
Trying a different way (it still has issues, though) to copy HTML files
in a more general manner. The issue with the current solution is, that
the "recent" version is copied to the top-level documentation, while
what is currently at `docs` is "moved" to the `next` folder. This cannot
be easily covered by the existing solution, as it requires not "dumb"
copy and paste, but requires to dive into the logic of Docusaurus.

Also deleting some left-over artifacts of previous license downloads.