docs: Back to skip HTML Links
Failed to link to HTML files (the downloaded licenses) in any feasible
manner. Instead, rolled back efforts and returned to the extra
disclaimer about issues with HTML files and not rendering any links.

Also, changed the license filenames to align with the spdx.io filenames
(a left-over from trying to trick the filenames not to end with
`*.html`).