docs(changelog): Link Release Notes
Using the new pattern, we would not have a `v2406.1.0-1` changelog entry
on `main`, but the changelog for `v2412.0.0-1` would contain all
changes, that were also published for `v2406.1.0-1`. Due to the "break"
we introduce with the new concept now, decided for a lightweight
approach, thus, only linking the previous changelog entries.