test: Enhance Contract Test
Now using facade to trigger cancellation as it should be.

Also, validating the status code, so that we may assume
a successful cancellation request.

Nevertheless, given our currently provided GCC sandbox
cancellation support within this test does not seem to
be have as expected anymore. We may need to decide
dropping the test eventually.