feat: Adding More Context
At least for the DownloadFromGlobalLinkAction we require at least
the result object, so that we can see, if we were able to identify
a PD submission ID.

Providing as much information to the adapt* method, that may be
relevant to decide and if to and how to adapt the retry delay.

This refactoring tries to provide only minimal impact on existing
implementations of `GlobalLinkAction`, which is why parameters,
results and such are exchanged again and again between the base
implementation and its extending implementations.