build: Fail on Warnings
At least in GCC Workspace, we have no extra control that prevents
warnings (especially deprecation warnings) to pass through unseen.

Now applying strict checks, so that we ensure not to oversee deprecation
warnings accidentally.

Required a little more fine grained control, because of the
false-positive warning:

    Supported source version 'RELEASE_8' from annotation processor
    'org.glassfish.corba.annotation.processing.ExceptionWrapperProcessor'
    less than -source '17'

Note, that this state fails due to (already existing) deprecation
warnings that need to be handled next.