docs: Add Since Tag
To denote from which version on a feature is available, adding
MDX Components `<Since/>` as well as (yet unused) `<Deprecated/>`.

Both use shields.io to render badges.

Decided for now, to make them float right. If we experience, that
this is not always applicable, we may consider to make this
configurable.