ci: Change GitHub Pages Deployment Trigger
We now expect, aligned with the current approval
process, that an approval of the newest CMCC version can
be detected by a merge from a `maintenance/*` branch to
`main`. We now take this as signal when to (automatically)
deploy the webpage from `main`.

If this misses a relevant update, we can still run the
job manually.