chore: Suppress False-Positive Warning
IntelliJ IDEA 2025.2.2 fails to determine nullability for generic
parameter usage incorrectly. Thus, if we explicitly state that
`<P extends @Nullable Object>` is a non-null type in the implementation,
IntelliJ IDEA fails to detect that. The JSpecify reference checker
correctly identifies this as valid.

Added `@SuppressWarnings` for now to the corresponding parameter.